+++ odoo/local-src/connector_esb/tests/test_export_saleorder.py
@@ -29,7 +29,7 @@ class ExportSaleOrderTestCase(SavepointCase):
         self.prod1.default_code = 'SKU01'
         self.prod1.default_code = 'SKU02'
         self.so1 = self.model.create({
-            'esb_ref': 'ref_123',
+            'esb_ref': 'ref_01',
             'partner_id': self.partner.id,
             'date_order': '2018-01-29',
             'sale_channel': 'fax',
@@ -45,8 +45,25 @@ class ExportSaleOrderTestCase(SavepointCase):
                     'product_uom_qty': 7,
                 })],
         })
+        self.so2 = self.model.create({
+            'esb_ref': 'ref_02',
+            'partner_id': self.partner.id,
+            'date_order': '2018-01-29',
+            'sale_channel': None,
+            'carrier_id': self.delivery.id,
+            'client_order_ref': 'whatever the client want',
+            'delivery_price': 23.5,
+            'suite_name': '0123434234',
+            'order_line': [
+                (0, 0, {
+                    'sequence': 1,
+                    'name': self.prod1.name,
+                    'product_id': self.prod1.id,
+                    'product_uom_qty': 7,
+                })],
+        })
 
-    def test_mapper(self):
+    def test_mapper_01(self):
         """ Generate data dict with mapper and check with what is expected """
         so = self.so1
         expected = {
@@ -54,6 +71,44 @@ class ExportSaleOrderTestCase(SavepointCase):
             'customer_id': so.partner_id.id,
             'date': so.date_order.split(' ')[0],
             'channel':  '03',  # the code for fax
+            'order_ref': so.client_order_ref,
+            'status': 'processing',
+            'shipping_method': so.carrier_id.esb_ref,
+            'apb_tax_amount': 0,
+            'order_amount': so.amount_total,
+            'tax_amount': int(so.amount_tax),
+            'shipping_amount': so.delivery_price,
+            'serial_no': int(so.suite_name),
+            'increment_id': so.esb_ref,
+            'lines': [{
+                'line_number': so.order_line[0].sequence,
+                'price': so.order_line[0].price_unit,
+                'price_inc_tax': (
+                    so.order_line[0].price_unit
+                    + so.order_line[0].price_reduce_taxinc
+                    - so.order_line[0].price_reduce
+                    ),
+                'qty_ordered': so.order_line[0].product_uom_qty,
+                'qty_delivered': so.order_line[0].qty_delivered,
+                'qty_cancelled': so.order_line[0].product_qty_canceled,
+                'qty_backorder': so.order_line[0].product_qty_unavailable,
+                'sku': so.order_line[0].product_id.default_code,
+                }]
+            }
+        with self.backend.work_on(self.model._name) as work:
+            mapper = work.component(usage='export.mapper')
+            values = mapper.map_record(so).values()
+        self.maxDiff = None
+        self.assertDictEqual(values, expected)
+
+    def test_mapper_02(self):
+        """Test default values."""
+        so = self.so2
+        expected = {
+            'erp_id': so.id,
+            'customer_id': so.partner_id.id,
+            'date': so.date_order.split(' ')[0],
+            'channel':  '01',  # Default is phone
             'order_ref': so.client_order_ref,
             'status': 'processing',
             'shipping_method': so.carrier_id.esb_ref,