+++ HISTORY.rst
@@ -37,6 +37,7 @@ Unreleased
 **Bugfixes**
 
 * ALCN-864: Fixing the csv document generated as a delivery note by improving float number format and adding no suite, lot name, tax amount.
+* ALCN-1420: Fix sale order export to esb: Verify prices default to 0, fix delivery carrier esb_ref, communication channel default to phone
 
 **Build**